// web/src/app/alarm/types/monitor.ts
export interface MetricItem {
    id: number;
    metric_group: number;
    metric_object: number;
    name: string;
    type: string;
    display_name?: string;
    display_description?: string;
    instance_id_keys?: string[];
    dimensions: any[];
    query?: string;
    unit?: string;
    displayType?: string;
    description?: string;
    viewData?: any[];
    style?: {
        width: string;
        height: string;
    };
    [key: string]: unknown;
}
// ...
export interface FiltersConfig {
    level: string[];
    state: string[];
    notify: string[];
    alarm_source: string[];
}